rename track-bundle files to track-reference
