+++ packages/core/src/track-reference/index.ts
@@ -1,2 +1,2 @@
-export * from './track-bundle.types';
-export * from './track-bundle.utils';
+export * from './track-reference.types';
+export * from './track-reference.utils';
