+++ packages/core/src/track-reference/track-reference.types.ts
